+++ httpolice/__init__.py
@@ -10,20 +10,12 @@ from httpolice.reports.text import text_report
 from httpolice.request import Request
 from httpolice.response import Response
 
-# Backward compatibility
-ERROR = Severity.error
-COMMENT = Severity.comment
-DEBUG = Severity.debug
-
 __all__ = [
     'Complaint',
     'Exchange',
     'Request',
     'Response',
     'Severity',
-    'COMMENT',
-    'DEBUG',
-    'ERROR',
     'check_exchange',
     'helpers',
     'html_report',