+++ CHANGELOG.rst
@@ -7,6 +7,16 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_
 (which means it is unstable until 1.0).
 
 
+Unreleased
+~~~~~~~~~~
+
+Removed
+-------
+- The deprecated constants
+  ``httpolice.ERROR``, ``httpolice.COMMENT``, ``httpolice.DEBUG``
+  have been removed. Use ``httpolice.Severity`` instead.
+
+
 0.3.0 - 2016-08-14
 ~~~~~~~~~~~~~~~~~~